+++ next.config.js
@@ -1,8 +1,11 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
-  output: 'export',
-  trailingSlash: true
+  images: {
+    domains: ["firebasestorage.googleapis.com"]
+  }
+  // output: 'export',
+  // trailingSlash: true
 }
 
 module.exports = nextConfig
