+++ package.json
@@ -16,27 +16,30 @@
     ]
   },
   "dependencies": {
+    "@ant-design/cssinjs": "^1.20.0",
     "@ant-design/icons": "^5.1.4",
+    "antd": "5.6.2",
+    "firebase": "9.22.2",
+    "next": "13.3.4",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "sharp": "^0.33.4",
+    "swiper": "^11.1.1",
+    "swr": "^2.2.5"
+  },
+  "devDependencies": {
     "@types/node": "20.3.1",
     "@types/react": "18.2.13",
     "@types/react-dom": "18.2.6",
     "@typescript-eslint/eslint-plugin": "5.60.0",
     "@typescript-eslint/parser": "5.60.0",
-    "antd": "5.6.2",
     "eslint": "8.43.0",
-    "eslint-config-next": "13.4.7",
+    "eslint-config-next": "13.3.4",
     "eslint-config-prettier": "8.8.0",
     "eslint-plugin-react": "7.32.2",
-    "firebase": "9.22.2",
-    "firebase-tools": "^12.4.6",
     "husky": "8.0.3",
     "lint-staged": "13.2.2",
-    "next": "13.4.7",
     "prettier": "2.8.8",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "swiper": "^11.1.1",
-    "swr": "^2.2.5",
     "typescript": "5.0.4"
   }
 }
